Refactoring: moving/removing files/folders. Added Row component as cubes wrapper
