+++ client/src/components/Mode/Shared/Desk/Cube/Cube.scss
