+++ client/src/components/Mode/Shared/Desk/Desk.scss
@@ -1,0 +1,19 @@
+@import './Row/Row.scss', './Cube/Cube.scss';
+
+.desk {
+    gap: 30px 0;
+    display: flex;
+    flex-direction: column;
+
+    .row {
+        &__header {
+            font-size: 30px;
+        }
+
+        &__cubes {
+            gap: 0 80px;
+            display: flex;
+            margin: 50px 0 25px;
+        }
+    }
+}
