--- client/src/components/Mode/Shared/Game/Desk/Desk.scss
@@ -1,12 +1,0 @@
-@import './Cubes/Cubes.scss';
-
-.desk {
-    display: flex;
-    flex-direction: column;
-
-    &__cubes {
-        display: flex;
-        flex-direction: column;
-        gap: 30px 0;
-    }
-}
