--- client/src/components/Mode/Shared/Game/Game.scss
@@ -1,5 +1,0 @@
-@import './Desk/Desk.scss', './Conclusion/Conclusion.scss';
-
-.game {
-    display: flex;
-}
