+++ client/src/components/Mode/Shared/Shared.scss
@@ -1,1 +1,1 @@
-@import './Game/Game.scss', './Status/Status.scss', './History/History.scss', './Players/Players.scss', './Countdown/Countdown.scss', './Controls/Controls.scss';
+@import './Status/Status.scss', './Desk/Desk.scss', './Conclusion/Conclusion.scss', './History/History.scss', './Players/Players.scss', './Countdown/Countdown.scss', './Controls/Controls.scss';
